requirements.txt
streamlit>=1.55.0
pymupdf>=1.27.0
anthropic>=0.86.0
openai>=2.29.0
tiktoken>=0.12.0
pandas>=2.0.0
numpy>=1.26.0
google-cloud-storage>=2.0.0
google-auth>=2.0.0
pyarrow>=14.0.0
